Move class LatestSnapshots to new file
