+++ java/statestore/src/test/java/sleeper/statestore/transactionlog/DynamoDBTransactionLogSnapshotMetadataStoreIT.java
@@ -29,7 +29,6 @@ import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.schema.Schema;
 import sleeper.dynamodb.tools.DynamoDBContainer;
-import sleeper.statestore.transactionlog.DynamoDBTransactionLogSnapshotMetadataStore.LatestSnapshots;
 
 import java.time.Instant;
 import java.util.List;
